Newsfeed Update
Created a "Post" class that contains author of post, target of post,
time, message, and icon.
User can create a new post and update the newsfeed.
Formatted the layout of newsfeed to use ListView
Created a custom adapter so that each entry in the ListView utilizes a
'post_box' layout
User needs at least 5 characters in the text box to post
Time shown on each post contains the difference in time from the current
time and the time the post was created eg. "4 days ago"