Send Message
Users can now send messages to each other. also added a refresh button
that refreshed the inbox and sent lists. can extend the function to
include course posts in the future. the only left on the Inbox Fragment
to do is to implement the checkbox multi select action bar functionality
for deleting messages